// scripts/src/kobis/types.ts
export interface Movie {
  // Original Korean: 영화명
  title: string;

  // Original Korean: 영화명(영문)
  englishTitle: string;

  // Original Korean: 제작연도
  productionYear: number;

  // Original Korean: 제작국가
  productionCountry: string;

  // Original Korean: 유형
  type: string;

  // Original Korean: 장르
  genre: string;

  // Original Korean: 제작상태
  productionStatus: string;

  // Original Korean: 감독
  director: string;

  // Original Korean: 제작사
  productionCompany: string;
}